Adds transpiled handlebars components
Adds a build task to generated handlebars components using meta-template
to convert the original nunjucks templates.

The build step works, in that it calls meta-template, but that fails to
convert some template features in to handlebars as there isn't an
equivilant feature there. This is the example thats failing:

```
{% if value == radio.value %} checked{% endif %}
```

Handlebars doesn't support comparisons in conditionals in the core
language (it can be added with plugins, but that won't be standard)